Restructuring into single modules
