--- src/py_back/enum/__init__.py
@@ -1,15 +1,0 @@
-"""
-Different enumerations types supported for python >= 3.8
-"""
-__all__ = [
-    "Enum",
-    "Flag",
-    "IntEnum",
-    "IntFlag",
-    "auto",
-    "unique",
-    "StrEnum",
-]
-from enum import Enum, Flag, auto, unique
-
-from ._types import IntEnum, IntFlag, StrEnum
